requirements.txt
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
chardet==5.0.0
charset-normalizer==2.1.1
discord.py==2.0.1
frozenlist==1.3.1
hiredis==2.0.0
idna==3.4
mpmath==1.2.1
multidict==6.0.2
Pillow==9.2.0
sympy==1.11.1
termcolor==2.0.1
typing_extensions==4.3.0
xmltodict==0.13.0
yarl==1.8.1
